+++ dakara_server/library/views.py
@@ -14,6 +14,7 @@ from library import models
 from library import serializers
 from library import permissions
 from library.query_language import QueryLanguageParser
+from library import views_feeder as feeder  # noqa F401
 
 
 UserModel = get_user_model()
@@ -161,6 +162,21 @@ class SongListView(ListCreateAPIViewWithQueryParsed):
             self.query_parsed = res
 
         return query_set.distinct().order_by(Lower("title"))
+
+    def get_serializer(self, *args, **kwargs):
+        """Return the serializer instance that should be used for validating and
+        deserializing input, and for serializing output.
+        """
+        data = kwargs.get("data")
+        many = kwargs.get("many")
+
+        # check if the serializer is used to deserialize data
+        # and check if the data is a list
+        if data and isinstance(data, list) and many is None:
+            return super().get_serializer(*args, many=True, **kwargs)
+
+        # otherwise
+        return super().get_serializer(*args, **kwargs)
 
 
 class SongView(RetrieveUpdateDestroyAPIView):